+++ main/pygame_xml_gui/src/classes/styleInserter.py
@@ -1,5 +1,7 @@
 import json
 
+from pygame_xml_gui.src.classes.errorHandler import ErrorHandler
+
 from ..files.paths import STYLE_LIGHT, STYLE_DARK
 from .widget import Widget
 
@@ -12,11 +14,13 @@ with open(STYLE_DARK, "r") as f:
     STYLES["dark"] = json.load(f)
 
 class StyleInserter:
-    def __init__(self, widgets: list[Widget]):
+    def __init__(self, widgets: list[Widget], classes: dict = None):
         self.__widgets = widgets
 
+        self.__classes: dict = classes
+
         self.__sanity_check()
-        self.__style: dict = STYLES[self.__widgets[0].attributes["pyStyle"]]
+        self.__canvas_style: dict = STYLES[self.__widgets[0].attributes["pyStyle"]]
 
         self.__run()
 
@@ -24,6 +28,8 @@ class StyleInserter:
         assert len(self.__widgets) == 1
         assert self.__widgets[0].name == "canvas"
         assert self.__widgets[0].attributes["pyStyle"] in STYLES.keys()
+        if self.__classes is not None:
+            assert isinstance(self.__classes, dict)
 
     def __run(self):
         self.__widgets = [
@@ -43,24 +49,30 @@ class StyleInserter:
             )
 
     def __get_widget_with_injected_style_attributes(self, widget: Widget) -> Widget:
-        pyStyle_attributes = self.__extract_style_attribute(widget)
+        pyStyle_attributes = self.__extract_pyStyle(widget)
+        pyClass_attributes = self.__extract_pyClass(widget)
 
-        attributes = widget.attributes
-        for k, v in self.__style.items():
-            if k not in attributes.keys():
-                attributes[k] = v
+        for k, v in self.__canvas_style.items():
+            widget.attributes[k] = v
+        
+        for k, v in pyClass_attributes.items():
+            widget.attributes[k] = v
+
         for k, v in pyStyle_attributes.items():
-            attributes[k] = v
-        attributes["anchor"] = "topleft"
+            widget.attributes[k] = v
 
-        return Widget(widget.name, attributes, widget.content)
+        # this line is necessary for the correct positioning of widgets
+        widget.attributes["anchor"] = "topleft"
 
-    def __extract_style_attribute(self, widget: Widget):
+        return Widget(widget.name, widget.attributes, widget.content)
+
+    def __extract_pyStyle(self, widget: Widget):
+        
         if "pyStyle" in widget.attributes.keys():
+            styles_dict = {}
             string = widget.attributes["pyStyle"]
             styles = string.split(";")
             styles = [s.strip() for s in styles if s.strip() != ""]
-            styles_dict = {}
             for style in styles:
                 try:
                     key = style.split("=")[0]
@@ -71,4 +83,24 @@ class StyleInserter:
                 except:
                     raise Exception(f"Could not interpret key-value pair: {style}")
             return styles_dict
+        
+        return {}
+    
+    def __extract_pyClass(self, widget: Widget) -> dict:
+
+        if "pyClass" in widget.attributes.keys() and self.__classes is not None:
+            styles_dict = {}
+            string = widget.attributes["pyClass"]
+            classes = string.split(" ")
+            classes = [s.strip() for s in classes if s.strip() != ""]
+            for class_ in classes:
+                if class_ not in self.__classes.keys():
+                    ErrorHandler.error(f"unknown class: {class_}", info=f"known classes: {list(self.__classes.keys())}")
+                try:
+                    for k, v in self.__classes[class_].items():
+                        styles_dict[k] = v
+                except Exception:
+                    ErrorHandler.error(f"Could not copy attributes of class '{class_}' into widget")
+            return styles_dict
+        
         return {}