+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## 0.2.2 (11.12.2023)
+# 0.2.2 (11.12.2023)
 - made tests more general
 - added short README.md
+- changed UI.set_structure to take a path to an xml file instead of the content of an xml file
+- added 'guides' folder
 
 ## 0.2.1 (11.12.2023)
 - improved set_pos method